deadline-sync/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
apscheduler>=3.10.0
httpx>=0.24.0
msal>=1.24.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.90.0
python-dateutil>=2.8.0
icalendar>=5.0.0
